refactor: add product card block
