+++ package.json
@@ -25,6 +25,7 @@
     "@types/jest": "^29.5.12",
     "@types/react": "^18.3.1",
     "@types/react-dom": "^18.3.0",
+    "@types/react-modal": "^3.16.3",
     "@types/react-redux": "^7.1.33",
     "@types/react-router": "^5.1.20",
     "@types/react-router-dom": "^5.3.3",
@@ -49,6 +50,7 @@
     "lint-staged": "^15.2.2",
     "prettier": "^3.2.5",
     "react-hook-form": "^7.51.4",
+    "react-modal": "^3.16.1",
     "react-test-renderer": "^18.3.1",
     "sass": "^1.76.0",
     "stylelint": "^16.5.0",
