+++ src/pages/ProductPage/ProductCardBlock/ProductCardBlock.module.scss
@@ -20,7 +20,7 @@
     }
 
     .productDescription {
-      @include montserrat(16px, 400, 20px);
+      @include montserrat(16px, 500, 20px);
     }
 
     .price {
