+++ src/pages/ProductPage/ProductSlider/productSlider.module.scss
@@ -20,5 +20,4 @@
     .productSlider {
       width: 330px;
     }
-
   }
